third_party/WebKit/LayoutTests/fast/dom/shadow/content-reprojection-recalc-expected.txt

CONSOLE WARNING: line 16: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Redistribution into same position should not cause style recalc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.internals is defined.
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS successfullyParsed is true

TEST COMPLETE
